# backend/docs/swagger.yaml
definitions:
  dtos_auth.ForgotPasswordDTO:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  dtos_auth.LoginDTO:
    properties:
      identifier:
        type: string
      password:
        type: string
    required:
    - identifier
    - password
    type: object
  dtos_auth.RegisterDTO:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  dtos_auth.ResetPasswordDTO:
    properties:
      email:
        type: string
      new_password:
        minLength: 8
        type: string
      otp:
        type: string
    required:
    - email
    - new_password
    - otp
    type: object
  dtos_auth.VerifyEmailDTO:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  dtos_roles.CreateRoleDTO:
    properties:
      role_name:
        type: string
      workspace_id:
        type: integer
    required:
    - role_name
    - workspace_id
    type: object
  dtos_roles.UpdateRoleDTO:
    properties:
      role_name:
        type: string
      uuid:
        type: string
    required:
    - role_name
    - uuid
    type: object
  dtos_workspace.AddMemberDTO:
    properties:
      role_uuid:
        type: string
      user_email:
        type: string
      user_uuid:
        type: string
      workspace_uuid:
        type: string
    required:
    - role_uuid
    - user_email
    - workspace_uuid
    type: object
  dtos_workspace.CreateWorkspaceDTO:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  models_auth.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
      username:
        type: string
      uuid:
        type: string
    type: object
  models_role.Role:
    properties:
      createdAt:
        type: string
      createdBy:
        $ref: '#/definitions/models_auth.User'
      createdByID:
        type: integer
      id:
        type: integer
      isSystem:
        type: boolean
      roleName:
        type: string
      updatedAt:
        type: string
      uuid:
        type: string
      workspaceID:
        type: integer
    type: object
info:
  contact: {}
paths:
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send OTP to email for password reset
      parameters:
      - description: Email
        in: body
        name: forgotPassword
        required: true
        schema:
          $ref: '#/definitions/dtos_auth.ForgotPasswordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Request password reset OTP
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with username/email and password
      parameters:
      - description: Login credentials
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/dtos_auth.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register new user
      parameters:
      - description: Register credentials
        in: body
        name: Register
        required: true
        schema:
          $ref: '#/definitions/dtos_auth.RegisterDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register user
      tags:
      - auth
  /auth/resend-verification:
    post:
      consumes:
      - application/json
      description: Resend email verification link
      parameters:
      - description: Email
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/dtos_auth.ForgotPasswordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Resend verification email
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password using OTP
      parameters:
      - description: Reset credentials
        in: body
        name: resetPassword
        required: true
        schema:
          $ref: '#/definitions/dtos_auth.ResetPasswordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset password with OTP
      tags:
      - auth
  /auth/verify-email:
    post:
      consumes:
      - application/json
      description: Verify user email with token
      parameters:
      - description: Verification token
        in: body
        name: verifyEmail
        required: true
        schema:
          $ref: '#/definitions/dtos_auth.VerifyEmailDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify email address
      tags:
      - auth
  /roles:
    get:
      description: Get all roles for a workspace
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models_role.Role'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Create a role inside a workspace
      parameters:
      - description: Create Role DTO
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/dtos_roles.CreateRoleDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new role
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: Update role name using role UUID
      parameters:
      - description: Update Role DTO
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/dtos_roles.UpdateRoleDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update role
      tags:
      - Roles
  /workspace:
    post:
      consumes:
      - application/json
      description: Create a workspace for the authenticated user
      parameters:
      - description: Workspace payload
        in: body
        name: workspace
        required: true
        schema:
          $ref: '#/definitions/dtos_workspace.CreateWorkspaceDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Workspace created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to create workspace
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new workspace
      tags:
      - Workspace
  /workspace/invite-member:
    post:
      consumes:
      - application/json
      description: Invite an existing user to join a workspace with a specific role.
        An invitation email will be sent to the user.
      parameters:
      - description: Add workspace member payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dtos_workspace.AddMemberDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Member invited successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request or resource not found
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: User already exists in workspace
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Invite a user to a workspace
      tags:
      - Workspace Members
swagger: "2.0"